requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
black==24.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
mysql-connector-python==8.3.0
python-dotenv==1.0.1
